chat history endpoint added
