+++ src/main/java/com/example/learninghub/chatbot/ChatMessageRepository.java
@@ -1,4 +1,4 @@
-package com.example.learninghub.chatmessage;
+package com.example.learninghub.chatbot;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
