--- src/main/java/com/example/learninghub/chatmessage/ChatMessageController.java
@@ -1,15 +1,0 @@
-package com.example.learninghub.chatmessage;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class ChatMessageController {
-
-    private final ChatMessageService chatMessageService;
-
-    @Autowired
-    public ChatMessageController(ChatMessageService chatMessageService) {
-        this.chatMessageService = chatMessageService;
-    }
-}
