--- src/main/java/com/example/learninghub/chatmessage/ChatMessageService.java
@@ -1,17 +1,0 @@
-package com.example.learninghub.chatmessage;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-@Service
-public class ChatMessageService {
-
-    private final ChatMessageRepository chatMessageRepository;
-
-    @Autowired
-    public ChatMessageService(ChatMessageRepository chatMessageRepository) {
-        this.chatMessageRepository = chatMessageRepository;
-    }
-
-
-}
